feat: init_hit_record.c & init_ray & get_point_ray...
tracing 디렉터리 내의 파일에 있던 함수들을 pixel디렉터리로 옮김.

+++ srcs/tracing/draw_pixel.c
@@ -22,8 +22,8 @@ void	draw_pixel(t_scene *scene, t_color3 color, double u, double v)
 	int	hex_color;
 
 	img = (int *)(scene->mlx_info.data_addr);
-	x_coord = u * (scene->canvas.width - 1);
-	y_coord = v * (scene->canvas.height - 1);
+	x_coord = u * (double)(scene->canvas.width - 1);
+	y_coord = v * (double)(scene->canvas.height - 1);
 	hex_color = get_hex_color(color);
 	img[(y_coord * scene->canvas.width) + x_coord] = hex_color;
 }
